requirements.txt
# 婴儿年龄预测项目依赖

# 核心依赖
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
PyYAML>=6.0
matplotlib>=3.5.0
seaborn>=0.11.0
joblib>=1.1.0

# 可选依赖
# jupyter>=1.0.0  # 用于notebook分析